refactor(core): 🔨 extract file utilities
