+++ Sources/DesktopManager/MonitorService.Display.cs
@@ -117,25 +117,6 @@ public partial class MonitorService {
         return string.Empty;
     }
 
-    private static string WriteStreamToTempFile(Stream stream) {
-        string path = Path.Combine(Path.GetTempPath(), Path.GetRandomFileName());
-        try {
-            using FileStream fs = File.Create(path);
-            stream.CopyTo(fs);
-            return path;
-        } catch {
-            DeleteTempFile(path);
-            throw;
-        }
-    }
-
-    private static void DeleteTempFile(string path) {
-        try {
-            File.Delete(path);
-        } catch (Exception ex) {
-            Console.WriteLine($"DeleteTempFile failed: {ex.Message}");
-        }
-    }
 
     private DesktopWallpaperPosition GetWallpaperPositionFallback() {
         try {
